Roles to copy/template files to server
The copy role copies folder recursively to the server. It also allows to
remove the previous folder contents before copying the new contents
over. The template role is alike except that it runs Jinja 2 templating
engine on files before copying.

There have been ideas to combine both roles in one. Currently there is
not enough time to do so.