Add config files for interface eth0/dns
These config files set the correct values for IPv6 and DNS resolution.
Under Linux DNS resolution works independent and is on the application-
layer in the OSI-Model. Therefore, I cannot set the DNS settings for a
specific IP version only.

I decided to set the DNS with IPv6 addresses and use the modern protocol
for resolution. This setting has no impact on the IP protocol used for
client (end-user browser) connections.

The file eth0 contains the manual IPv6 settings for the eth0 interface.
The hosting provider does not set the values correctly, that's why we
set the values manually in this config file.